// frontend-dealers/src/app/(authenticated)/car-evaluation/_components/ExteriorSection/_components/Front/types.ts
export type Condition =
  | "Good"
  | "Scratched"
  | "Dented"
  | "Damaged"
  | "Repaired"
  | "Replaced"
  | "Rusted"
  | "Repainted"
  | "Not Available"
  | "Welded"
  | "Spots"
  | "Cracked"
  | "Non-Functional"
  | "Not Working"
  | "Non-Functional"
  | "Clamped"
  | "Chipped"
  | "Faded"
  | "Damaged"
  | "Not Available"
  | "Foggy"
  | "Working"
  | "Other";